Change DrawingSizeChangedMessage value to float.

+++ SketchOverlay/ViewModels/DrawingToolWindowViewModel.cs
@@ -76,7 +76,7 @@ public partial class DrawingToolWindowViewModel : ObservableObject
             else
                 _selectedDrawingSize = value;
 
-            WeakReferenceMessenger.Default.Send(new DrawingSizeChangedMessage(_selectedDrawingSize));
+            WeakReferenceMessenger.Default.Send(new DrawingSizeChangedMessage((float)_selectedDrawingSize));
         }
     }
